pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tlog"
version = "1.0.0"
description = "Amateur Radio Logbook - система учета радиосвязей"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Django>=5.2",
    "Pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "python-dotenv>=1.0.1",
    "gunicorn>=23.0.0",
    "whitenoise>=6.8.2",
    "celery>=5.4.0",
    "redis>=5.3.0",
    "django-extensions>=3.2.3",
    "django-debug-toolbar>=4.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-django>=4.9.0",
    "factory-boy>=3.3.1",
    "requests>=2.32.3",
    "argon2-cffi>=23.1.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "myproject.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = ["-v", "--tb=short"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"